Refactor and new library geekclub/sound #8
