+++ examples/sound2.py
@@ -1,31 +1,15 @@
 import sys
 sys.path.append('..')
 from geekclub.pyscratch import *
-import time
-
-# Install simpleaudio with: pip3 install simpleaudio
-import simpleaudio as sa
+from geekclub.sound import *
 
 create_canvas()
 
 # Not really necessary for this example, but why not:
 sprite = ImageSprite('images/face.gif')
 sprite.centre()
-
-# Get free to use sounds here: https://freesound.org/
-def load_sound(file):
-    return sa.WaveObject.from_wave_file(file)
-
-bpm = 3 * 60
-def bps(): return bpm / 60
-def beat_ms():
-    return int(1000 / bps())
-
-print(bpm, bps(), beat_ms())
-
-def rest(beats):
-    time.sleep(beat_ms() * beats / 1000)
-               
+   
+set_bpm(180)
 laser = load_sound('sounds/laser.wav')
 drum = load_sound('sounds/bass-drum.wav')
 hh = load_sound('sounds/hh-cymbal.wav')
